PANW/Prisma_Cortex_Cloud_toolkit_v1.2/requirements.txt
# Prisma Cloud Toolkit v1.2 - Security Hardened Dependencies
# All versions pinned for security and reproducibility

# Core dependencies
requests==2.31.0
pyyaml==6.0.1
cryptography==41.0.7

# Security and validation
pydantic==2.5.0
jsonschema==4.20.0

# Logging and monitoring
structlog==23.2.0

# Retry logic and resilience
tenacity==8.2.3

# Optional: For enhanced security features
# python-dotenv==1.0.0  # For environment variable management
# hvac==1.2.0  # For HashiCorp Vault integration (optional)
# boto3==1.34.0  # For AWS Secrets Manager (optional)
# azure-identity==1.15.0  # For Azure Key Vault (optional)
# azure-keyvault-secrets==4.8.0  # For Azure Key Vault (optional)

# Testing (optional, for development)
# pytest==7.4.3
# pytest-cov==4.1.0
# pytest-mock==3.12.0
